backend/requirements.txt
fastapi[standard]
pydantic
pymongo
passlib
starlette
motor
fastapi_mail
PyJWT
python-dotenv
streamlit
faker
langchain
langchain-openai
langchain-ollama
boto3
azure-storage-blob